fix(tasks): validate Kanban identity and restore line limits
