+++ src/renderer/src/components/automations/automation-target-availability.ts
@@ -1,5 +1,5 @@
 import type { Automation } from '../../../../shared/automations-types'
-import { getRepoExecutionHostId, parseExecutionHostId } from '../../../../shared/execution-host'
+import { parseExecutionHostId } from '../../../../shared/execution-host'
 import {
   describeRuntimeCompatBlock,
   evaluateRuntimeCompat
@@ -16,6 +16,11 @@ import type { ProjectHostSetup } from '../../../../shared/project-types'
 import type { Repo } from '../../../../shared/repo-types'
 import type { Worktree } from '../../../../shared/worktree/types'
 import type { TaskSourceHostAvailability } from '../task-source-context-summary'
+import {
+  getAutomationSshTargetId,
+  repoHostMatchesRunContext,
+  setupHostMatchesRunContext
+} from './automation-target-availability-host-matching'
 
 export type AutomationTargetAvailability =
   | {
@@ -151,40 +156,6 @@ export function getAutomationTargetAvailability({
   }
 }
 
-function getRuntimeTargetHostId(target: AutomationHostTarget | null | undefined): string | null {
-  return target?.kind === 'environment'
-    ? `runtime:${encodeURIComponent(target.environmentId)}`
-    : null
-}
-
-function setupHostMatchesRunContext(
-  setupHostId: string,
-  runHostId: string,
-  target: AutomationHostTarget | null | undefined
-): boolean {
-  if (setupHostId === runHostId) {
-    return true
-  }
-  const targetHostId = getRuntimeTargetHostId(target)
-  // Why: remote-runtime project lists project the server-local host as runtime:<env>,
-  // while CLI-created automations can preserve the server's durable local run host.
-  return targetHostId !== null && setupHostId === targetHostId && runHostId === 'local'
-}
-
-function repoHostMatchesRunContext(
-  repo: Repo,
-  runHostId: string,
-  target: AutomationHostTarget | null | undefined
-): boolean {
-  if (runHostId === getRepoExecutionHostId(repo)) {
-    return true
-  }
-  const targetHostId = getRuntimeTargetHostId(target)
-  // Why: repos fetched from a remote runtime are owned by runtime:<env> in the
-  // renderer, but saved automations still target the host setup that runs there.
-  return targetHostId !== null && getRepoExecutionHostId(repo) === targetHostId
-}
-
 function getAutomationSourceAvailability(
   sourceContext: TaskSourceContext | null | undefined,
   sourceHostAvailability: readonly TaskSourceHostAvailability[] | undefined
@@ -303,17 +274,6 @@ export function getRuntimeAutomationAvailability(
     return unavailable('runtime-update-required', describeRuntimeCompatBlock(compat))
   }
   return { canRunNow: true, reason: 'available', message: null }
-}
-
-function getAutomationSshTargetId(automation: Automation, repo: Repo): string | null {
-  const parsedHost = parseExecutionHostId(automation.runContext?.hostId)
-  if (parsedHost?.kind === 'ssh') {
-    return parsedHost.targetId
-  }
-  if (automation.executionTargetType === 'ssh' && automation.executionTargetId.trim()) {
-    return automation.executionTargetId
-  }
-  return repo.connectionId?.trim() || null
 }
 
 function unavailable(
