+++ src/renderer/src/components/sidebar/folder-workspace-composer-submit-kanban.test.ts
@@ -1,0 +1,124 @@
+// @vitest-environment happy-dom
+
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
+import type { FolderWorkspace } from '../../../../shared/folder-workspace-types'
+import type { ProjectGroup } from '../../../../shared/project-group-types'
+import type * as NewWorkspaceModule from '@/lib/new-workspace'
+
+const mocks = vi.hoisted(() => ({
+  activateAndRevealFolderWorkspace: vi.fn(),
+  ensureAgentStartupInTerminal: vi.fn()
+}))
+
+vi.mock('@/lib/worktree-activation', async (importOriginal) => {
+  const actual = await importOriginal<Record<string, unknown>>()
+  return { ...actual, activateAndRevealFolderWorkspace: mocks.activateAndRevealFolderWorkspace }
+})
+
+vi.mock('@/lib/new-workspace', async (importOriginal) => {
+  const actual = await importOriginal<typeof NewWorkspaceModule>()
+  return {
+    ...actual,
+    ensureAgentStartupInTerminal: mocks.ensureAgentStartupInTerminal
+  }
+})
+
+import { submitFolderWorkspaceCreate } from './folder-workspace-composer-submit'
+
+function makeProjectGroup(): ProjectGroup {
+  return {
+    id: 'group-1',
+    name: 'Platform',
+    parentPath: '/repo/platform',
+    parentGroupId: null,
+    createdFrom: 'folder-scan',
+    tabOrder: 0,
+    isCollapsed: false,
+    color: null,
+    createdAt: 1,
+    updatedAt: 1
+  }
+}
+
+function makeFolderWorkspace(overrides: Partial<FolderWorkspace> = {}): FolderWorkspace {
+  return {
+    id: 'folder-workspace-1',
+    projectGroupId: 'group-1',
+    name: 'hi',
+    folderPath: '/repo/platform/hi',
+    linkedTask: null,
+    comment: '',
+    isArchived: false,
+    isUnread: false,
+    isPinned: false,
+    sortOrder: 0,
+    lastActivityAt: 1,
+    createdAt: 1,
+    updatedAt: 1,
+    ...overrides
+  }
+}
+
+describe('submitFolderWorkspaceCreate kanban', () => {
+  beforeEach(() => {
+    mocks.activateAndRevealFolderWorkspace.mockReturnValue({ primaryTabId: 'tab-1' })
+    Object.assign(window, {
+      api: {
+        agentTrust: {
+          markTrusted: vi.fn().mockResolvedValue(undefined)
+        }
+      }
+    })
+  })
+
+  afterEach(() => {
+    mocks.activateAndRevealFolderWorkspace.mockReset()
+    mocks.ensureAgentStartupInTerminal.mockReset()
+    Reflect.deleteProperty(window, 'api')
+    vi.restoreAllMocks()
+  })
+
+  it('creates a Kanban folder workspace with its bound source context', async () => {
+    const createFolderWorkspace = vi.fn(async () => makeFolderWorkspace())
+    const linkedWorkItem = {
+      provider: 'kanban' as const,
+      type: 'issue' as const,
+      number: 0,
+      title: '4123 Fix checkout retry',
+      url: 'https://kanban.fpimi.ru/?task=4123',
+      kanbanIdentifier: '4123'
+    }
+    const linkedTaskSourceContext = {
+      kind: 'task-source' as const,
+      provider: 'kanban' as const,
+      projectId: 'group-1',
+      hostId: 'local' as const,
+      providerIdentity: {
+        provider: 'kanban' as const,
+        serverUrl: 'https://kanban.fpimi.ru' as const
+      }
+    }
+
+    await submitFolderWorkspaceCreate({
+      projectGroup: makeProjectGroup(),
+      name: '',
+      lastAutoName: '',
+      linkedWorkItem,
+      linkedTaskSourceContext,
+      note: '',
+      quickAgent: null,
+      autoRenameBranchFromWork: true,
+      agentCmdOverrides: {},
+      createFolderWorkspace,
+      onOpenChange: vi.fn()
+    })
+
+    expect(createFolderWorkspace).toHaveBeenCalledWith({
+      projectGroupId: 'group-1',
+      name: '4123 Fix checkout retry',
+      connectionId: null,
+      linkedTask: linkedWorkItem,
+      linkedTaskSourceContext
+    })
+  })
+})
